ci: Change the GitHub Actions trigger, drop Celery 3 support (#261)
* ci: Change the GitHub Actions trigger

Now, it will run tests for every pull request made to the `master` branch, as well as when merges are pushed to `master`.

* Drop support for Celery 3

Celery 3 depends on `use_2to3` which is no longer supported by any of the latest patch versions of CPython and Pypy, for >= 3.7.

The CI for Celery 3 was not able to run on any Python provided by GitHub Actions. It is time to fully decommission it.